code health: extract the Future type to future.go
This is not a critical change. It just splits request.go into
request.go and future.go.
